probe: Mark discarded probe samples in-line
This is a follow-up to @richardjm's introduction of SAMPLES_DISCARD_FIRST.

With discarding enabled, the sample logging creates misleading visual "sample
groups" that don't in fact correspond to samples for the same probe operation.
This PR corrects that issue by marking discarded samples on the same line as
their probe values.

Signed-off-by: Garth Snyder <garth@garthsnyder.com>